fix: classroom_id url error + update error handling
fix: users will be redirected to the login page if their classroom/id url is accessed in another browser where they are not authenticated.

prior to fix:  if a user tried to access their classroom/id url in another browser it would throw a currentUser undefined error.

feat: updated error handling on data fetch requests. Users will now be notified in their UI if there is an error when fetching their data and will be prompted to click a link to redirect them back to the dashboard or homepage so that they can try again.