fix: user can't use back button on dashboard page
fix: users can now use the back button on the dashboard page when signing up/in to return back to their previous page before the sign up or sign in page.

prior to fix: when a user would press back whilst on the dashboard, it would try and redirect them back to the sign up or sign in page and then redirect them back to the dashboard

In order to replace the sign up/sign in page in the browser history stack, I refactored the sign up/sign in page to use 'redirect()' instead of 'router.push()'. This is because by default in client components, redirect will replace the current URL in the browser history stack.

This fix/refactor creates a better UX and improves the users ability to navigate around the app.