fix: error from unauthenticated user accessing democlass page
fix: if an unathenticated user tries to access the democlass page through the url pathname (/democlass) it will now not throw an error and instead redirect them to the sign in page.

prior to fix: an error would be thrown stating that the currentUser.uid was null.

I implemented the optional chaining operator (?.) on the currentUser.uid in the components where the error was being thrown.